Refactor: Split semantic module into cohesive sub-modules
- Split `src/semantic/mod.rs` (4600+ lines) into focused modules:
  - `control_flow.rs`: Loops, conditionals, match expressions
  - `declarations.rs`: Type, trait, and function definitions
  - `expressions.rs`: Expression analysis and assembler feeding
  - `patterns.rs`: Struct instantiation and trait method calls
  - `conversion.rs`: Assembly to HIR lowering logic
- Keep `src/semantic/mod.rs` as an orchestrator and facade
- Update imports to use the new structure
- Verify with `cargo check` and `cargo test`
- Document architectural change in `.jules/atlas.md`
